mcp_common_requirements.txt
fastapi
uvicorn
httpx
requests
python-dotenv
pydantic
pydantic-settings
openai>=1.0.0
